+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Release v0.1.0
-Initial xmipp4-compute-cuda release. Still work in progress.
+Initial xmipp4-hardware-cuda release. Still work in progress.
 
-- Up to date compute interface implementation with the CUDA backend
+- Up to date hardware interface implementation with the CUDA backend
 - Efficient caching memory allocator based on PyTorch